requirements.txt
python-telegram-bot==20.7
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1 